+++ packages/11ty/_includes/components/figure/audio/element.js
@@ -2,6 +2,7 @@ import { html } from '#lib/common-tags/index.js'
 import chalkFactory from '#lib/chalk/index.js'
 
 const logger = chalkFactory('Figure Video')
+
 /**
  * Renders an embedded soundcloud audio player
  *
@@ -17,7 +18,7 @@ const logger = chalkFactory('Figure Video')
 export default function(eleventyConfig) {
   const figureMediaEmbedUrl = eleventyConfig.getFilter('figureMediaEmbedUrl')
   const audioElements = {
-    soundcloud({ id, mediaId, mediaType }) {
+    soundcloud({ id, mediaId, mediaType, lazyLoading }) {
       if (!mediaId) {
         logger.error(`Cannot render SoundCloud component without 'media_id'. Check that figures data for id: ${id} has a valid 'media_id'`)
         return ''
@@ -29,6 +30,7 @@ export default function(eleventyConfig) {
         <iframe
           allow="autoplay"
           frameborder="no"
+          loading="${ lazyLoading ?? 'lazy' }"
           height="166"
           scrolling="no"
           src="${embedUrl}"
@@ -37,7 +39,7 @@ export default function(eleventyConfig) {
       `
     }
   }
-  return function ({ id, mediaId, mediaType }) {
-    return audioElements[mediaType]({ id, mediaId, mediaType })
+  return function ({ id, mediaId, mediaType, lazyLoading }) {
+    return audioElements[mediaType]({ id, mediaId, mediaType, lazyLoading })
   }
 }