+++ packages/11ty/_includes/web-components/lightbox/index.js
@@ -1,32 +1,56 @@
-import { LitElement, html } from 'lit'
+import { LitElement, html, render, unsafeCSS } from 'lit'
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js'
 
 /**
- * Lightbox lit-element web component
+ * Lightbox lit-element web component markup and browser runtime
  *
- * This component renders image slides with navigation UI.
- * It provides a default slot for passing markup to render:
+ * This component renders a custom element (q-lightbox) with slots for 
+ * figures data, styles, and ui. On startup it loads its data, inserts its
+ * slides' markup, and manages a few internal properties and attributes during
+ * the component lifecycle.
  *
- * To display an element as a slide, provide it with a
- * `data-lightbox-slide` attribute set to any value
- * `data-lightbox-slide-id` attribute set to a unique id string
+ * Each slide (see `_lightboxSlide`) carries a
+ * `q-lightbox-slides__slide` class
+ * `id` attribute set to the figure's id
+ * `data-lightbox-current` (optional, any value) attribute lays out and displays the slide 
+ * `data-lightbox-preload` (optional, any value) attribute lays out the slide without display
  *
- * This lightbox provides access to controls with the following data attributes:
+ * This lightbox expects its UI slot to have elements for a few controls:
  * - `data-lightbox-fullscreen` triggers fullscreen on click and indicates status
  * - `data-lightbox-next` triggers rendering of next slide on click
  * - `data-lightbox-previous` triggers rendering of previous slide on click
  *
- * It dynamically updates the content of elements with these data attributes:
+ * It dynamically updates the content of its UI labels with these data attributes:
  * - `data-lightbox-counter-current` displays the current slide index
  * - `data-lightbox-counter-total` displays total number of slides
+ * 
+ * For each figure's data:
+ * - Aligned to an internal data representation that eases slide logic and markup
+ * - Inserted into slide markup with the appropriate slot and added to the lightbox
  */
 class Lightbox extends LitElement {
+
   static properties = {
     currentId: { attribute: 'current-id', type: String },
+    preloadNextId: { attribute: 'preload-next-id', type: String },
+    preloadPrevId: { attribute: 'preload-prev-id', type: String },
+  }
+
+  setupStyles() {
+    let styleElement = this.querySelector('style[slot="lightbox-styles"]')
+
+    if (styleElement && styleElement.sheet) {
+      this.adoptedStyleSheets = [ styleElement.sheet ]    
+    }
+
   }
 
   constructor() {
     super()
+
+    // NB: UI controls check this.slides so load figures first
+    this.setupStyles()
+    this.setupFiguresData()
     this.setupFullscreenButton()
     this.setupNavigationButtons()
     this.setupKeyboardControls()
@@ -53,13 +77,13 @@ class Lightbox extends LitElement {
   }
 
   get slides() {
-    return this.querySelectorAll('[data-lightbox-slide]')
+    return this.querySelectorAll('.q-lightbox-slides__slide')
   }
 
   get slideIds() {
     return Array
       .from(this.slides)
-      .map((slide) => slide.dataset.lightboxSlideId)
+      .map((slide) => slide.id)
   }
 
   get currentSlide() {
@@ -121,6 +145,87 @@ class Lightbox extends LitElement {
     this.currentId = previousId
   }
 
+  /**
+   * @function _alignSlideData
+   * @param {Object} figure - data for a quire figure
+   * @return {Object} - data w/ a few useful props added
+   */
+  _alignSlideData(figure) {
+    const {
+      aspect_ratio: aspectRatio='widescreen',
+      caption,
+      credit,
+      id,
+      isSequence,
+      label,
+      mediaType
+    } = figure
+
+    const isAudio = mediaType === 'soundcloud'
+    const isVideo = mediaType === 'video' || mediaType === 'vimeo' || mediaType === 'youtube'
+
+    return { ...figure, isAudio, isVideo, aspectRatio }
+  }
+
+  /**
+   * @function _lightboxSlide
+   * @param {Object} figure - slide data
+   * @return {Lit.html} - Lit HTML component
+   * 
+   * Composites slide data into the slide markup and injects pre-serialized figure elements
+   **/
+  _lightboxSlide(figure) {
+    const { id, 
+      mediaType, 
+      figureElementContent,
+      annotationsElementContent, 
+      aspectRatio, 
+      isVideo,
+      isAudio,
+      label, 
+      labelHtml, 
+      caption, 
+      captionHtml, 
+      credit, 
+      creditHtml 
+    } = figure
+
+    const slideMediaClass = `q-lightbox-slides__element--${mediaType}`
+    const videoClass = isVideo ? `q-lightbox-slides__element--video q-lightbox-slides__element--${aspectRatio}` : ''
+    const audioClass = isAudio ? 'q-lightbox-slides__element--audio' : ''
+
+    return html`<div class="q-lightbox-slides__slide" slot="slides" id="${id}">
+      <div class="q-lightbox-slides__element ${slideMediaClass} ${videoClass} ${audioClass}">
+        ${unsafeHTML(figureElementContent)}
+      </div>
+      <div class="q-figure-slides__slide-ui">
+        <div class="q-lightbox-slides__caption ${ label || caption || credit ? '' : 'is-hidden' }">
+          <span class="q-lightbox-slides__caption-label ${ label ? '' : 'is-hidden' }">${unsafeHTML(labelHtml)}</span>
+          <span class="q-lightbox-slides__caption-content ${ caption || credit ? '' : 'is-hidden' }">${captionHtml ? unsafeHTML(captionHtml) : ''} ${creditHtml ? unsafeHTML(creditHtml) : ''}</span>
+        </div>
+        ${unsafeHTML(annotationsElementContent ?? '')}
+      </div>
+    </div>`
+  }
+
+  setupFiguresData() {
+    const figuresData = this.querySelector('script[type="application/json"]')
+    const figures = JSON.parse(figuresData.innerHTML)
+    this.figures = figures
+
+    figures.forEach( fig => {
+      console.log('Rendering',fig.id)
+      const dat = this._alignSlideData(fig)
+      
+      // NB: This is done in lightDOM *not* shadowDOM
+      let slideElement = document.createElement('template') 
+      render(this._lightboxSlide(dat),slideElement.content)
+
+      const slide = slideElement.content.cloneNode(true)
+      this.appendChild(slide)
+    })
+  }
+
   setupFullscreenButton() {
     if (!this.fullscreenButton) return
 
@@ -178,9 +283,18 @@ class Lightbox extends LitElement {
     if (!this.currentSlide) return
 
     this.slides.forEach((slide) => {
-      if (slide.dataset.lightboxSlideId !== this.currentId)
+      if (slide.id !== this.currentId) {
         delete slide.dataset.lightboxCurrent
+      }
+
+      if ([this.preloadNextId,this.preloadPrevId].includes(slide.id)) {
+        slide.dataset.lightboxPreload = true
+      } else {
+        delete slide.dataset.lightboxPreload
+      }
+
     })
+
     this.currentSlide.dataset.lightboxCurrent = true
   }
 
@@ -191,12 +305,20 @@ class Lightbox extends LitElement {
   render() {
     if (!this.slides.length) return ''
     this.currentId = this.slideIds[this.currentSlideIndex]
+    this.preloadNextId = this.currentSlideIndex + 1 < this.slideIds.length ? this.slideIds[this.currentSlideIndex + 1] : undefined
+    this.preloadPrevId = this.currentSlideIndex - 1 >= 0 ? this.slideIds[this.currentSlideIndex - 1] : undefined
     this.updateCurrentSlideElement()
     this.updateCounterElements()
 
+    // NB: `lightbox-styles` are dynamically loaded by the constructor
     return html`
       <div class="q-lightbox">
-        <slot></slot>
+        <slot name="lightbox-styles"></slot>
+        <slot name="data"></slot>
+        <div class="q-lightbox-slides">
+          <slot name="slides"></slot>
+        </div>
+        <slot name="ui"></slot>
       </div>
     `
   }