+++ packages/11ty/_plugins/collections/index.js
@@ -3,7 +3,7 @@ import sortCollection from './sort/index.js'
 
 /**
  * Add Collections and Apply Transforms
- * 
+ *
  * Nota bene: The Eleventy API does not make collections data accessible
  * from the plugin context. Adding `collections` and `transforms` sequentially
  * in the same file allows access to `collections` data from `transforms`
@@ -19,14 +19,15 @@ export default function(eleventyConfig, options = {}) {
    * Add sorted "all" collection
    */
   eleventyConfig.addCollection('allSorted', function (collectionApi) {
-    return collectionApi.getAll().sort(sortCollection)
+    return collectionApi.getAllSorted().sort(sortCollection)
   })
 
   /**
    * Add eleventy-generated collections to collections object
    */
   eleventyConfig.addCollection('temp', function (collectionApi) {
-    const eleventyCollections = collectionApi.getAll()[0].data.collections
+    const allCollections = collectionApi.getAll()
+    const eleventyCollections = allCollections.length > 0 ? collectionApi.getAll()[0].data.collections : {}
     Object.assign(collections, eleventyCollections)
     return []
   })
@@ -37,7 +38,7 @@ export default function(eleventyConfig, options = {}) {
   for (const name in filters) {
     eleventyConfig.addCollection(name, function (collectionApi) {
       collections[name] = collectionApi
-        .getAll()
+        .getAllSorted()
         .filter(filters[name])
         .sort(sortCollection)
       return collections[name]